Merge network loop into main loop
This removes the need for one of the channels. I think I like it better
this way as the code stands, though it will probably make sense to break
it back out if I add the option for another screen UI.

Also fix sending output to the socket when connected, which got
accidentally dropped in the async change.