Handle subtask early exits
The network listener task in particular is susceptible to failing due to
a conflict of binding; the others seem much less likely to fail, but we
may as well treat them all uniformly.

The new task wrapper type is the nicest way I could think of to handle
the logic we want: the main loop polls the tasks for doneness, but we
also want to also poll all of them afterward to shut down gracefully and
log errors. To avoid duplicating code for handling output values or
having to make the handles `Option`s, the wrapper separates checking
whether the inner task has completed from extracting the result.